# Reduce allocs in NewAnonymousUser