--- a/user.go
+++ b/user.go
@@ -50,9 +50,12 @@ func NewUser(key string) User {
 	return User{Key: &key}
 }
 
+var (
+	anonymous = true
+)
+
 // NewAnonymousUser creates a new anonymous user identified by the given key.
 func NewAnonymousUser(key string) User {
-	anonymous := true
 	return User{Key: &key, Anonymous: &anonymous}
 }