@eli-darkly 2019-10-03
> A downside is that because User.Anonymous is public the value of the global anonymous flag can be changed.

That's true... and unfortunately I think that outweighs the slight performance gain of this change. Since Go doesn't have such a thing as a reference to an immutable value, there's simply no safe way to reuse a pointer like this.

Ideally, we would not have defined this field as a `*bool` in the first place, since there is no functional difference between leaving it unset and setting it to `false` as far as LaunchDarkly's user indexing behavior is concerned. It is a pointer for historical reasons, because originally we were serializing this struct directly to JSON and wanted to omit the property entirely unless it was true. We are now stuck with that at least until the next major version, for backward compatibility. In the meantime I don't think we should make a change that has such a clear potential for problems.